Update README, various clean-ups
Week switching was broken in my app. But was working in the demo app running locally.
Spent hours trying to figure it out.
Eventually realised that the sample code in the README was missing the
`sync` modifier on the props passed to the `Kalendar` component.

Add this to the README with another recommendation to use the demo app
as the starting point as this would have saved me a lot of time.

Also while trying to figure this out I made some cleanups:\
 - kalendar-container:
   - removed some unused code
   - removed `previousDay`, `previousWeek` etc methods and replaced them
   with one `changeDay` method
 - kalendar-event - make `portal-target` self closing for brevity
 - kalendar-weekview: using `v-for` and `v-if` together is not
 recommended -
 https://vuejs.org/v2/guide/conditional.html#v-if-with-v-for - used a
 computed property instead
 - added `Vue.prototype.$kalendar = {}` to the module `install` method
 - serve-dev - removed the `generate-shadow` class - not used anywhere
 - ran `npm audit` updated some dependencies